backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
websockets==11.0.3
python-dotenv==1.0.0
httpx==0.24.1

openai==1.3.0
